Add option to include full domain name for js, css, and image files.
This adds a "host_name" config variable to the main settings.

When set to a domain name, it will prepend JS and CSS tag sources with the specified domain name. It will also add the domain name to images specified in CSS files.

By specifying this option with an environment variable, assets can be accessed locally for development environments and accessed through a CDN in production environments.

Signed-off-by: Blake Harvey <bharvey@sittercity.com>